Shooting cannons' projectiles move right (x coord).

+++ src/net/johnbrooks/fjg/drawables/tower/Tower.java
@@ -46,7 +46,7 @@ public abstract class Tower
 
     public void draw()
     {
-        Draw.drawTexture(baseTexture, (int)x, (int)y, width, height);
+        Draw.drawTexture(baseTexture, x, y, width, height);
     }
 
     protected abstract void shoot();
